+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 4.0.0-BETA3 (YYYY-MM-DD)
 
+### Enhancements
+
+* [ObjectServer] `SyncUserInfo` now also exposes a users metadata using `SyncUserInfo.getMetadata()`
+
 ### Internal
 
 * Upgraded to Realm Sync 2.0.0-rc16.